Register all tests in tests/ as CTest tests
Each .cpp file in a subdirectory of tests is now registered as a CTest test that will be run as part of `make test`.

Also installs tests as part of `make install`.

Note: libEl has to be installed first before `make test` can be run. There's probably a way to run the tests in-place, which is more idiomatic, but I couldn't figure out how to build the test binary as part of the CTest test. Instead, I've adapted the sandbox/Makefile to allow make to be used as a test driver.